Moved the stats overview classes to the correct package.
